predabst: Farkas coefficients may be negative, so use int64 rather than unsigned.

+++ src/muz/predabst/farkas_util.h
@@ -38,11 +38,11 @@ struct lambda_info {
     }
 };
 
-expr_ref make_linear_combination(vector<unsigned> const& coeffs, expr_ref_vector const& inequalities);
+expr_ref make_linear_combination(vector<int64> const& coeffs, expr_ref_vector const& inequalities);
 
 bool mk_exists_forall_farkas(expr_ref const& fml, expr_ref_vector const& vars, expr_ref_vector& constraints, vector<lambda_info>& lambda_infos, bool eliminate_unsat_disjuncts = false);
 
-bool get_farkas_coeffs(expr_ref_vector const& assertions, vector<unsigned>& coeffs);
+bool get_farkas_coeffs(expr_ref_vector const& assertions, vector<int64>& coeffs);
 
 void well_founded_bound_and_decrease(expr_ref_vector const& vsws, expr_ref& bound, expr_ref& decrease);
